Make it work

+++ secrets/generate-standup-kubeconfig-secret.sh
@@ -16,27 +16,40 @@ function cleanup() {
 
 trap cleanup EXIT
 
-# for provider in aws azure kvm; do
-for provider in aws; do
+for provider in aws azure kvm; do
     installation=${provider_installations[$provider]}
     echo "creating $provider kubeconfig for $installation"
 
+    # Create a temp kubeconfig for the installation
     installation_kubeconfig="$tmp_dir/tmp-kubeconfig"
     KUBECONFIG=$installation_kubeconfig opsctl create kubeconfig -i $installation --certificate-common-name-prefix test-infra --ttl 30 > /dev/null
 
+    # Get the installation server
     server=$(KUBECONFIG=$installation_kubeconfig kubectl config view -o jsonpath='{.clusters[].cluster.server}')
+
+    # Get the name of the secret from the CP where the service account token and cert are stored
     secret_name=$(KUBECONFIG=$installation_kubeconfig kubectl -n giantswarm get serviceaccount $service_account -o jsonpath='{.secrets[].name}')
+    if [[ -z "$secret_name" ]]; then
+        echo "Could not get the secret name of $service_account service account on $installation. You can create everything you need on the CP by using the files in the control-plane folder."
+        exit 1
+    fi
+
+    # Read service account token from the secret
     token=$(KUBECONFIG=$installation_kubeconfig kubectl -n giantswarm get secret/"$secret_name" -o jsonpath='{.data.token}' | base64 --decode)
 
-    KUBECONFIG=$installation_kubeconfig kubectl -n giantswarm get secret/"$secret_name" -o jsonpath='{.data.ca\.crt}' > "$tmp_dir/ca_cert"
+    # Read service account cert from the secret
+    KUBECONFIG=$installation_kubeconfig kubectl -n giantswarm get secret/"$secret_name" -o jsonpath='{.data.ca\.crt}'  | base64 --decode > "$tmp_dir/ca_cert"
 
+    # Remove the temp kubeconfig
     rm "$installation_kubeconfig"
 
+    # Add data from the service account to the result kubeconfig
     touch "$result_kubeconfig"
     KUBECONFIG="$result_kubeconfig" kubectl config set-cluster "giantswarm-cluster-$installation" --server="$server" --certificate-authority="$tmp_dir/ca_cert" --embed-certs=true
 
-    KUBECONFIG="$result_kubeconfig" kubectl config set-context "giantswarm-cluster-$installation" --cluster "giantswarm-cluster-$installation" --user "giantswarm-user-$installation"
     KUBECONFIG="$result_kubeconfig" kubectl config set-credentials "giantswarm-user-$installation" --token="$token"
+    KUBECONFIG="$result_kubeconfig" kubectl config set-context "giantswarm-$installation" --cluster "giantswarm-cluster-$installation" --user "giantswarm-user-$installation"
+    KUBECONFIG="$result_kubeconfig" kubectl config use-context "giantswarm-$installation"
 done
 
 kubectl create secret generic standup-kubeconfig -n test-workloads --from-file=kubeconfig="$result_kubeconfig" --dry-run=client -o yaml > "$script_dir/standup-kubeconfig-secret.yaml"
